Eco refs

See merge request almukhan/dashboard!82

+++ database/migrations/2020_10_01_114136_create_eco_refs_discont_coef_bars_table.php
@@ -23,7 +23,7 @@ class CreateEcoRefsDiscontCoefBarsTable extends Migration
             $table->float('barr_coef',8,2);
             $table->float('discont',8,2);
             $table->float('oil_cost',8,2);
-            $table->integer('macro');
+            $table->float('macro', 8,2);
             $table->timestamps();
         });
     }
